requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication & Security
Werkzeug==3.0.1
WTForms==3.1.1
bcrypt==4.3.0

# LLM Providers
openai==1.3.7
anthropic==0.7.8
google-cloud-aiplatform==1.35.0
google-cloud-language==2.10.1
google-cloud-translate==3.12.0

# Document Extraction & AI
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-core==2.14.0

# Text Processing & Analysis
spacy==3.7.2
nltk==3.8.1
langdetect==1.0.9
textstat==0.7.8
beautifulsoup4==4.13.4
python-docx==1.2.0
pypdf==5.1.0
pdfplumber==0.11.4
lxml==6.0.0

# ML & Embeddings
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
pyarrow==16.1.0

# Graph & Visualization
networkx==3.2.1
rdflib==7.0.0
owlrl==6.0.2

# Web & HTTP
requests==2.31.0
urllib3==2.1.0
httpx==0.27.2
Jinja2==3.1.2

# Development & Utilities
python-dotenv==1.0.0
click==8.1.7
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1

# File handling
Pillow==11.3.0
python-magic==0.4.27

# Configuration & Logging
PyYAML==6.0.1
structlog==23.2.0

# Testing (optional but recommended)
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2

# OED Parser dependency
langextract
